// src/enums/columns.enum.ts
export enum Columns {
  Select = 'select',
  CodeClient = 'CodeClient',
  DateDocument = 'DateDocument',
  Nature = 'Nature',
  Statut = 'Statut',
  Nommage = 'Nommage',
  TypeDocument = 'TypeDocument',
  NomClient = 'NomClient',
  NumFacture = 'NumFacture',
  NomSociete = 'NomSociete',
  SousCategory = 'SousCategory',
  DateFacture = 'DateFacture',
  Options = 'Options',
  DatesConseil = 'DatesConseil',
  DateDebutFacture = 'DateDebutFacture',
  Periodicite = 'Periodicite',
  Montant = 'Montant',
  DateFinFacture = 'DateFinFacture',
  Duplicata = 'Duplicata',
  Acheteur = 'Acheteur',
  AcheteurPending = 'AcheteurPending',
  Reference = 'Reference',
  Client = 'Client',
  ClientFacture = 'ClientFacture',
  Createur = 'Createur',
  DernierIntervenant = 'Dernier intervenant',
  DirectionRegionale = 'DirectionRegionale',
  Fournisseur = 'Fournisseur',
  DateSignature = 'DateSignature',
  DatesATraiter = 'DatesATraiter',
  DateFinValidite = 'DateFinValidite',
  Partenaire = 'Partenaire',
  Famille = 'Famille',
  SousFamilleMR = 'SousFamilleMR',
  FamilleMR = 'FamilleMR',
  ThemeMR = 'ThemeMR',
  DateCreation = 'DateCreation',
  VisibilityClient = 'VisibilityClient',
  Titre = 'Titre',
  LastModification = 'LastModification',
  ClassementMR = 'ClassementMR',
  NomClientMR = 'NomClientMR',
  CodeClientMR = 'CodeClientMR',
  NommageMR = 'NommageMR',
  TitreMR = 'TitreMR',
  Origine = 'Origine',
  OptionsMR = 'OptionsMR',
  OptionsHomeMR = 'OptionsHomeMR',
  CategorieProduit = 'CategorieProduit',
  CategorieProduitMixed = 'CategorieProduitMixed',
  OptionsConseil = 'OptionsConseil',
  Produit = 'Produit',
  Description = 'Description',
  Name = 'Name',
  Date = 'Date',
  NumeroDossier = 'NumeroDossier',
  DomaineActivite = 'DomaineActivite',
  Metier = 'Metier',
  NiveauExpertise = 'NiveauExpertise',
  Preview = 'Preview',
  AssocieGerance = 'Associe',
  AssocieGerancePending = 'Associe_Traitement',
  PartenaireGerance = 'PartenaireGerance',
  TypeDocumentGerance = 'TypeDocumentGerance',
  SousDossiers = 'SousDossiers',
  DatesGerance = 'DatesGerance',
  OptionsGerance = 'OptionsGerance',
  CategorieProduitGerance = 'CategorieProduitGerance',
  ProduitGerance = 'ProduitGerance',
  FamilleGeranceAssocies = 'FamilleGerance',
  NommageGerance = 'NommageGerance',
  Titre_AP = 'Titre_AP',
  Nature_AP = 'Nature_AP',
  Annee_AP = 'Annee_AP',
  Dossier_AP = 'Dossier_AP',
  Localisation_AP = 'Localisation_AP',
  Thematique_AP = 'Thematique_AP',
  Remarques_AP = 'Remarques_AP',
  Date_AP = 'Date_AP',
  Options_AP = 'Options_AP',
  AcheteurSofiral = 'AcheteurSofiral',
  TitreSofiral = 'Titre_Sofiral',
  OptionsSofiral = 'Options_Sofiral',
}